+++ backend/application/client/client_repository.py
@@ -1,6 +1,6 @@
 import abc
 
-from application.requests import ClientResponse
+from application.responses import ClientResponse
 from domain.client import Client
