+++ backend/infrastructure/mongo/client_repository.py
@@ -1,7 +1,7 @@
 from bson import ObjectId
 
 from application.client.client_repository import AbstractClientRepository
-from application.requests import ClientResponse
+from application.responses import ClientResponse
 from domain.client import Client
 from domain.exceptions import ClientNotFoundError
 from infrastructure.mongo.mongo_client import MongoDBClient
@@ -13,7 +13,9 @@ class ClientRepositoryMongo(AbstractClientRepository):
 
     def register_client_db(self, client: Client) -> ClientResponse:
         client_data = client.model_dump()
-        client_data["id"] = str(self.client_collection.insert_one(client_data).inserted_id)
+        client_data["id"] = str(
+            self.client_collection.insert_one(client_data).inserted_id
+        )
         return ClientResponse(**client_data)
 
     def get_client_db(self, client_id: str) -> ClientResponse:
