+++ backend/infrastructure/mongo/product_repository.py
@@ -1,7 +1,7 @@
 from bson import ObjectId
 
 from application.product.product_repository import AbstractProductRepository
-from application.requests import ProductResponse
+from application.responses import ProductResponse
 from domain.exceptions import ProductNotFoundError
 from domain.product import Product
 from infrastructure.mongo.mongo_client import MongoDBClient
